simple_api/test_app/serializers.py
from django.conf import settings
from django.contrib.auth import get_user_model
from rest_framework import serializers

from .models import Lesson

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = get_user_model()
        fields = ('id', 'username', 'first_name', 'last_name')


class LessonSerializer(serializers.ModelSerializer):
    class Meta:
        model = Lesson
        fields = '__all__'